% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runParaApp}
\alias{runParaApp}
\title{Start the Survey Solutions Paradata Viewer Application}
\usage{
runParaApp(
  launch.browser = TRUE,
  mapwidget.option = c("leaflet", "mapdeck"),
  useTrackingPass = NULL,
  trackServer = NULL,
  trackServerPort = 5432,
  trackServerDB = NULL,
  trackServerUser = NULL,
  trackServerPass = NULL,
  trackServerTable = NULL
)
}
\arguments{
\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only. The value of this parameter can also be a
function to call with the application's URL.}

\item{mapwidget.option}{Selection of map at start-up, mapdeck or leaflet, defaults to leaflet if NULL}

\item{useTrackingPass}{Passcode to access tracking section, if NULL, tracking is disabled}

\item{trackServer}{PostgreSQL tracking DB URL}

\item{trackServerPort}{PostgreSQL tracking DB port}

\item{trackServerDB}{PostgreSQL tracking DB name}

\item{trackServerUser}{PostgreSQL tracking DB user}

\item{trackServerPass}{PostgreSQL tracking DB pasword}

\item{trackServerTable}{PostgreSQL tracking DB table with tracking data}
}
\description{
A wrappter function to start the application. Please make sure you have read the
documentation on how to use the app.
}
\details{
This application is part of the large set of tools, to facilitate survey implementation with
\href{https://docs.mysurvey.solutions/}{Survey Solutions}. Survey Solutions Paradata is data about the actual
data collection process, however it comes with some challenges in processing. This application seeks to minimize these challenges,
by transforming the data, and subsequently producing plots and reports from it.
}
\examples{
\dontrun{
# Start app in the current working directory
runApp()

# Start app in a subdirectory called myapp
runApp("myapp")
}

## Only run this example in interactive R sessions
if (interactive()) {
  options(device.ask.default = FALSE)

  # Apps can be run without a server.r and ui.r file
  runApp(list(
    ui = bootstrapPage(
      numericInput('n', 'Number of obs', 100),
      plotOutput('plot')
    ),
    server = function(input, output) {
      output$plot <- renderPlot({ hist(runif(input$n)) })
    }
  ))


  # Running a Shiny app object
  app <- shinyApp(
    ui = bootstrapPage(
      numericInput('n', 'Number of obs', 100),
      plotOutput('plot')
    ),
    server = function(input, output) {
      output$plot <- renderPlot({ hist(runif(input$n)) })
    }
  )
  runApp(app)
}
}
